Common methods moved to NotificationHelper
