+++ Api/HappyTravel.Edo.Api.csproj
@@ -18,7 +18,7 @@
 
   <ItemGroup>
     <PackageReference Include="AspNetCore.HealthChecks.Redis" Version="3.1.2" />
-    <PackageReference Include="CSharpFunctionalExtensions" Version="2.13.2" />
+    <PackageReference Include="CSharpFunctionalExtensions" Version="2.15.1" />
     <PackageReference Include="CsvHelper" Version="25.0.0" />
     <PackageReference Include="Elasticsearch.Net" Version="7.10.0" />
     <PackageReference Include="FloxDc.CacheFlow" Version="1.7.0" />
@@ -65,7 +65,6 @@
     <ProjectReference Include="..\HappyTravel.Edo.Common\HappyTravel.Edo.Common.csproj" />
     <ProjectReference Include="..\HappyTravel.Edo.Data\HappyTravel.Edo.Data.csproj" />
     <ProjectReference Include="..\HappyTravel.Edo.NotificationCenter\HappyTravel.Edo.NotificationCenter.csproj" />
-    <ProjectReference Include="..\HappyTravel.Edo.Notifications\HappyTravel.Edo.Notifications.csproj" />
   </ItemGroup>
 
   <ItemGroup>
