+++ Api/Services/Notifications/INotificationOptionsService.cs
@@ -1,8 +1,8 @@
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
 using HappyTravel.Edo.Api.Models.Agents;
-using HappyTravel.Edo.Api.Models.Notifications;
 using HappyTravel.Edo.Notifications.Enums;
+using HappyTravel.Edo.Notifications.Models;
 
 namespace HappyTravel.Edo.Api.Services.Notifications
 {
