+++ HappyTravel.Edo.Notifications/HappyTravel.Edo.Notifications.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="CSharpFunctionalExtensions" Version="2.15.1" />
+  </ItemGroup>
+
 </Project>
