+++ HappyTravel.Edo.Notifications/Models/SlimNotificationOption.cs
@@ -1,6 +1,6 @@
 using HappyTravel.Edo.Notifications.Enums;
 
-namespace HappyTravel.Edo.Api.Models.Notifications
+namespace HappyTravel.Edo.Notifications.Models
 {
     public readonly struct SlimNotificationOptions
     {
